Remove more not used code and fix tests for "core"
